+++ tests/helpers/index.test.js
@@ -4,3 +4,4 @@ import './noop.test';
 import './number-utils.test';
 import './string-utils.test';
 import './object-utils.test';
+import './log-message.test';