+++ tests/scriptlets/trusted-set-cookie.test.js
@@ -112,9 +112,11 @@ test('Set cookie with invalid expires', (assert) => {
         if (input.indexOf('trace') > -1) {
             return;
         }
-        assert.strictEqual(input,
+        assert.strictEqual(
+            input,
             `${name}: Invalid offsetExpiresSec value: ${expiresSec}`,
-            'logs correctly on invalid offsetExpiresSec');
+            'logs correctly on invalid offsetExpiresSec',
+        );
     };
     runScriptlet(name, [cName, cValue, `${expiresSec}`]);