+++ src/helpers/compatibility-redirects.js
@@ -82,6 +82,7 @@ const redirects = [
     },
     {
         adg: 'google-ima3',
+        ubo: 'google-ima.js',
     },
     {
         adg: 'gemius',